+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ NEXT (YYYY-MM-DD)
 
 * Additional fields in ``Legend`` class (https://github.com/weaveworks/grafanalib/pull/25)
 * SingleStat panel class added (https://github.com/weaveworks/grafanalib/pull/22)
+* Additional fields in ``XAxis`` class (https://github.com/weaveworks/grafanalib/pull/28)
 
 
 0.1.2 (2017-01-02)